requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
musicbrainzngs==0.7.1
oauthlib==3.2.2
python-dateutil==2.8.2
python3-discogs-client==2.6
requests==2.28.2
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.2
zipp==3.9.0
